… esphome/components/web_server/server_index.h