This PR adds a number of fixes, small improvements and refinements to web site UI:

Numbers:
 - Use step value to define accuracy (value of 1.000000 while step is 1 looked weird)
 - Add unit of measurement into state
 - Handle NaN number values with a more user friendly NA
 - Fix https://github.com/esphome/esphome-webserver/issues/11

Sensors:
 - Handle NaN sensor values with a more user friendly NA (e.g. was 'nan °C', became 'NA')

Climate:
 - Fix behavior of mode labels: clicking on a label activates the proper mode instead of always the first one
 - Minor cleanup of html code (removed useless arguments) 
 
Plus PR adds a helper function to derive accuracy in decimals from an increment step. Previous was a part of ClimateTraits.

Before:
![image](https://user-images.githubusercontent.com/48646998/184499742-fb816d73-9e00-4613-86d9-a5a6e5981248.png)

After:
![image](https://user-images.githubusercontent.com/48646998/184499764-eab2cbe9-62ca-4f33-8e0a-005e6f85ea0e.png)

Uses web server content changes from this PR https://github.com/esphome/esphome-webserver/pull/12

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Other

**Related issue or feature (if applicable):** fixes (https://github.com/esphome/esphome-webserver/issues/11)

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** esphome/esphome-docs#<esphome-docs PR number goes here>

## Test Environment

- [ ] ESP32
- [ ] ESP32 IDF
- [x] ESP8266

## Example entry for `config.yaml`:
<!--
  Supplying a configuration snippet, makes it easier for a maintainer to test
  your PR. Furthermore, for new integrations, it gives an impression of how
  the configuration would look like.
  Note: Remove this section if this PR does not have an example entry.
-->

```yaml
# Example config.yaml
climate:
  - platform: bang_bang
    name: ${friendly_devicename} Heating Controller
    sensor: ${temperature_sensor}
    default_target_temperature_low: ${target_temperture_low} °C
    default_target_temperature_high: ${target_temperture_high} °C
    id: lytko_therm
    visual:
      min_temperature: ${min_temperature} °C
      max_temperature: ${max_temperature} °C
      temperature_step: ${temperature_step} °C

    heat_action:
      - switch.turn_on: heater
      - light.turn_on: blue_led

    idle_action:
      - switch.turn_off: heater
      - light.turn_off: blue_led
  - platform: thermostat
    name: "Thermostat Climate Cooler Controller"
    sensor: ${temperature_sensor}
    default_target_temperature_high: 30.5 °C
    min_cooling_off_time: 3s
    min_cooling_run_time: 3s
    min_idle_time: 1s
    visual:
      min_temperature: 30.0 °C
      max_temperature: 31.0 °C
      temperature_step: 0.05°C
    cool_action:
      - light.turn_on: blue_led
    idle_action:
      - light.turn_off: blue_led

  - platform: pid
    name: "PID Climate Controller"
    sensor: ${temperature_sensor}
    default_target_temperature: 21°C
    heat_output: pwm_output
    visual:
      min_temperature: ${min_temperature} °C
      max_temperature: ${max_temperature} °C
      temperature_step: ${temperature_step} °C
    control_parameters:
      kp: 0.49460
      ki: 0.00487
      kd: 12.56301


  - platform: thermostat
    name: "Thermostat Climate Controller"
    sensor: ${temperature_sensor}
    default_target_temperature_low: 20 °C
    default_target_temperature_high: 22 °C
    min_cooling_off_time: 3s
    min_cooling_run_time: 3s
    min_heating_off_time: 3s
    min_heating_run_time: 3s
    min_idle_time: 3s
    cool_action:
      - light.turn_on: blue_led
    heat_action:
      - light.turn_on: blue_led
    idle_action:
      - light.turn_on: blue_led

  - platform: thermostat
    name: "Thermostat Climate Heating Controller"
    sensor: ${temperature_sensor}
    default_target_temperature_low: 0.0 °C
    min_heating_off_time: 3s
    min_heating_run_time: 3s
    min_idle_time: 1s
    heat_action:
      - light.turn_on: blue_led
    idle_action:
      - light.turn_off: blue_led
    visual:
      min_temperature: -1000.0 °C
      max_temperature: 1000.0 °C
      temperature_step: 100 °C

  - platform: template
    name: ${friendly_devicename} Locked Screen Brightness
    id: locked_brightness
    min_value: 0
    max_value: 100
    step: 1
    optimistic: true
    restore_value: true
    entity_category: config
    unit_of_measurement: '%'
    mode: slider

  - platform: template
    name: qq
    min_value: 0
    max_value: 1000
    step: 1
    optimistic: true
    restore_value: true
    entity_category: config
    unit_of_measurement: '%'
    mode: slider

  - platform: template
    name: ${friendly_devicename} Locked Screen Brightness 2
    min_value: 30
    max_value: 31
    step: 0.05
    optimistic: true
    restore_value: true
    entity_category: config
    unit_of_measurement: '%'
    mode: slider

  - platform: template
    name: ${friendly_devicename} Locked Screen Brightness 3
    min_value: -1000
    max_value: 1000
    step: 100
    optimistic: true
    restore_value: true
    entity_category: config
    unit_of_measurement: '%'
    mode: slider

  - platform: template
    name: ${friendly_devicename} Locked Screen Brightness 4
    min_value: ${min_temperature}
    max_value: ${max_temperature}
    step: ${temperature_step}
    optimistic: true
    restore_value: true
    entity_category: config
    unit_of_measurement: "°C"
    mode: slider
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).